+++ src/main/java/com/greghaskins/spectrum/Spectrum.java
@@ -6,9 +6,9 @@ import static com.greghaskins.spectrum.internal.hooks.BeforeHook.before;
 import static com.greghaskins.spectrum.model.BlockConfiguration.Factory.focus;
 import static com.greghaskins.spectrum.model.BlockConfiguration.Factory.ignore;
 
-import com.greghaskins.spectrum.internal.ConstructorBlock;
-import com.greghaskins.spectrum.internal.IdempotentBlock;
 import com.greghaskins.spectrum.internal.Suite;
+import com.greghaskins.spectrum.internal.blocks.ConstructorBlock;
+import com.greghaskins.spectrum.internal.blocks.IdempotentBlock;
 import com.greghaskins.spectrum.internal.hooks.Hook;
 import com.greghaskins.spectrum.internal.hooks.HookContext;
 import com.greghaskins.spectrum.internal.hooks.LetHook;
