+++ src/main/java/com/greghaskins/spectrum/internal/Spec.java
@@ -1,5 +1,7 @@
 package com.greghaskins.spectrum.internal;
 
+import com.greghaskins.spectrum.internal.blocks.NotifyingBlock;
+
 import org.junit.runner.Description;
 import org.junit.runner.notification.RunNotifier;
