+++ src/main/java/com/greghaskins/spectrum/internal/Suite.java
@@ -1,6 +1,7 @@
 package com.greghaskins.spectrum.internal;
 
 import com.greghaskins.spectrum.Block;
+import com.greghaskins.spectrum.internal.blocks.NotifyingBlock;
 import com.greghaskins.spectrum.internal.hooks.HookContext;
 import com.greghaskins.spectrum.internal.hooks.Hooks;
 import com.greghaskins.spectrum.model.BlockConfiguration;
