+++ src/main/java/com/greghaskins/spectrum/internal/blocks/ConstructorBlock.java
@@ -1,4 +1,4 @@
-package com.greghaskins.spectrum.internal;
+package com.greghaskins.spectrum.internal.blocks;
 
 import com.greghaskins.spectrum.Block;
